requirements.txt
scvi-tools
scanpy
matplotlib
seaborn
scikit-learn
pandas
numpy

